style: 修改目录结构
- 将Api分级，改为server/auth...
- 将Rpc分级，改为rpc/user...
- rpc中增加sekeleton与xxxstub，区分服务器和客户端代码
- rpc中将原types修改为xxxrpc，减少命名冲突
