+++ README.md
@@ -16,6 +16,8 @@
 ### 目录
 
 ```
+├─cmd
+│  └─initctl
 ├─configs
 ├─docs
 │  ├─devel
@@ -24,31 +26,38 @@
 │  │  ├─en-US
 │  │  └─zh-CN
 │  │      ├─api
-│  │      └─apiserver
+│  │      ├─apiserver
+│  │      ├─arpcclient
+│  │      └─arpcserver
 │  └─images
 ├─internal
-│  ├─authserver
-│  │  ├─api
-│  │  │  └─login
-│  │  ├─apply
-│  │  ├─config
-│  │  ├─reply
-│  │  └─svc
 │  ├─pkg
 │  │  ├─apply
 │  │  └─reply
-│  └─userrpc
-│      ├─handler
-│      ├─service
-│      ├─svc
-│      ├─types
-│      └─userclient
+│  ├─rpc
+│  │  ├─order
+│  │  └─user
+│  │      ├─skeleton
+│  │      │  ├─handler
+│  │      │  └─service
+│  │      ├─svc
+│  │      ├─userclient
+│  │      ├─userrpc
+│  │      └─userstub
+│  └─server
+│      ├─auth
+│      │  ├─api
+│      │  │  └─login
+│      │  ├─apply
+│      │  ├─config
+│      │  ├─reply
+│      │  └─svc
+│      └─order
 │-pkg
 │   ├─apiserver
 │   ├─arpcclient
 │   ├─arpcserver
-│   └─etcdx
-│-gitignore
+│   └─etcdx│-gitignore
 │-CHANGELOG
 │-go.mod
 │-Makefile
