+++ go.mod
@@ -9,5 +9,4 @@ require (
 	go.etcd.io/etcd/api/v3 v3.5.1
 	go.etcd.io/etcd/client/v3 v3.5.1
 	gopkg.in/yaml.v2 v2.3.0
-	honnef.co/go/tools v0.2.2
 )
