+++ internal/rpc/user/userrpc/user.go
@@ -1,4 +1,4 @@
-package types
+package userrpc
 
 import "context"
